modified dist file and src file
